Se guarda informacion del video al subir el archivo

+++ src/main/java/personal/netflix/clone/app/ServicesImpl/FileServiceImpl.java
@@ -1,32 +1,49 @@
 package personal.netflix.clone.app.ServicesImpl;
 
-import org.springframework.http.codec.multipart.FilePart;
+import com.xuggle.xuggler.IContainer;
+import org.springframework.beans.factory.annotation.Autowired;
+
 import org.springframework.stereotype.Service;
+import personal.netflix.clone.app.entities.VideoInfo;
 import personal.netflix.clone.app.services.FileService;
-import reactor.core.publisher.Mono;
 
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.Collection;
+import java.util.concurrent.TimeUnit;
 
 @Service
 public class FileServiceImpl implements FileService {
 
-    private static final Path VIDEO_PATH = Paths.get("./src/main/resources/videos");
-    public Boolean saveVideo(Mono<FilePart> filePartMono){
-        System.out.println("Path: " + VIDEO_PATH.toString());
-        filePartMono
-                .doOnNext(fp -> System.out.println("Received File : " + fp.filename()))
-                .flatMap(fp -> fp.transferTo(VIDEO_PATH.resolve(fp.filename())))
-                .then();
-        //TODO Crear validaciones para comprobar si existe previamente.
-        return true;
+    public static final Path VIDEO_PATH = Paths.get("./src/main/resources/videos");
+
+    @Autowired
+    private VideoInfoServiceImpl videoInfoService;
+
+    private void createVideoInfo(String fileName, VideoInfo videoInfo){
+        IContainer container = IContainer.make();
+        container.open(VIDEO_PATH + "/" + fileName, IContainer.Type.READ, null);
+        videoInfo.setDuration(TimeUnit.MICROSECONDS.toSeconds(container.getDuration()));
+        videoInfo.setSize(container.getFileSize());
+        videoInfo.setMediaType(fileName.split("\\.")[1]);
+        videoInfo.setBitRate(container.getBitRate());
+        videoInfo.setFileName(fileName);
+        container.close();
     }
-    public Boolean deleteVideo(String name){
+
+
+    public void saveVideo(String fileName,VideoInfo videoInfo){
+        createVideoInfo(fileName,videoInfo);
+        videoInfoService.saveOrUpdateVideoInfo(videoInfo);
+    }
+    public Boolean deleteVideoByName(String name){
         boolean result = false;
         try {
-            result = Files.deleteIfExists(Paths.get(VIDEO_PATH + "/" + name));
+            if(videoInfoService.deteleVideoInfoByName(name)){
+                result = Files.deleteIfExists(Paths.get(VIDEO_PATH + "/" + name));
+            }
         } catch (IOException e) {
             e.printStackTrace();
         }
